Trait proj::Coord

pub trait Coord<T>where
    T: CoordinateType,{
    fn x(&self) -> T;
    fn y(&self) -> T;
    fn from_xy(x: T, y: T) -> Self;
}

A point in two dimensional space. The primary unit of input/output for proj.

By default, any numeric (x, y) tuple implements Coord, but you can conform your type to Coord to pass it directly into proj.

See the geo-types feature for interop with the geo-types crate

Required Methods

fn x(&self) -> T

fn y(&self) -> T

fn from_xy(x: T, y: T) -> Self